+++ tests/parse/test_parse_clnsig.py
@@ -158,15 +158,19 @@ def test_parse_complex_clnsig(cyvcf2_variant):
 
 
 def test_parse_clnsig_transcripts(cyvcf2_variant):
-    ## GIVEN a variant with classic clinvar annotations
-    transcripts = [{"clnsig": ["likely_benign"]}]
+    ## GIVEN a variant with slash-separated values or values that start with underscore
+    transcripts = [
+        {"clnsig": ["pathogenic/likely_pathogenic", "likely_pathogenic", "pathogenic", "_other"]}
+    ]
 
     ## WHEN parsing the annotations
     clnsig_annotations = parse_clnsig(cyvcf2_variant, transcripts=transcripts)
 
     ## THEN assert that they where parsed correct
-    assert len(clnsig_annotations) == 1
-    assert clnsig_annotations[0]["value"] == "likely_benign"
+    assert len(clnsig_annotations) == 3
+    for clnsig in ["pathogenic", "likely_pathogenic", "other"]:
+        clnsig_dict = {"value": clnsig}
+        assert clnsig_dict in clnsig_annotations
 
 
 def test_is_pathogenic_pathogenic(cyvcf2_variant):