+++ CHANGELOG.md
@@ -6,9 +6,10 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
 ## [x.x.x]
 ### Added
+- CNV report link in cancer case side navigation
 ### Fixed
 - missing `vcf_cancer_sv` and `vcf_cancer_sv_research` to manual.
-
+- Split ClinVar multiple clnsig values (slash-separated) and strip them of underscore for annotations without accession number
 ### Changed
 - Do not freeze mkdocs-material to version 4.6.1
 - Remove pre-commit dependency
@@ -18,7 +19,6 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ### Added
 - Editable cases comments
 - Editable variants comments
-- CNV report link in cancer case side navigation
 ### Fixed
 - Empty variant activity panel
 - STRs variants popover